+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.16.0
 
 - **FEAT**: New `ShadTimePicker` and `ShadDatePickerFormField` components.
-- **FIX**: `maxLength` and `maxLengthEnforcement` not working on `ShadInput`
+- **FIX**: `maxLength`, `maxLengthEnforcement` and `showCursor` not working on `ShadInput`
 - **CHORE**: Set minimum Flutter version to `3.24.0`
 - **CHORE**: Remove `trackColor` from `ShadSwitch` (thanks to @RaghavTheGreat)